+++ doc/about/changes.md
@@ -32,6 +32,8 @@ in progress
   in the same directory as the resulting PCH
 - The `install_name` for libraries on macOS is now always an absolute path,
   instead of using `@rpath`-based paths prior to installation
+- MSBuild backend can now build libraries with no source files that link to a
+  whole-archive library
 
 ---